edit inductor test and performance test upload to cloudwatch
