+++ test/sagemaker_tests/pytorch/training/integration/sagemaker/test_distributed_operations_inductor.py
@@ -67,7 +67,7 @@ def can_run_smmodelparallel_efa(ecr_image):
 @pytest.mark.skip_test_in_region
 @pytest.mark.skip_inductor_test
 def test_dist_operations_cpu(framework_version, ecr_image, sagemaker_regions, instance_type, dist_cpu_backend):
-    instance_type = instance_type or 'ml.c4.xlarge'
+    instance_type = instance_type or 'ml.p2.xlarge'
     function_args = {
             'framework_version': framework_version,
             'instance_type': instance_type,
