--- test/sagemaker_tests/pytorch/training/integration/sagemaker/test_inductor_helper.py
@@ -1,53 +1,0 @@
-import boto3
-from botocore.config import Config
-from botocore.exceptions import ClientError
-import datetime
-import pandas as pd
-import os
-import argparse
-import logging
-
-LOGGER = logging.getLogger(__name__)
-
-AWS_DEFAULT_REGION = "us-west-2"
-
-def get_boto3_session(region=AWS_DEFAULT_REGION):
-    """Get boto3 session with us-east-1 as default region used to connect to AWS services."""
-    return boto3.session.Session(region_name=region)
-
-def get_cloudwatch_client(region=AWS_DEFAULT_REGION):
-    """Get AWS CloudWatch client object. Currently assume region is IAD (us-east-1)"""
-    return get_boto3_session(region=region).client("cloudwatch")
-
-def put_metric_data(metric_name, namespace, unit, value, dimensions):
-    """Puts data points to cloudwatch metrics"""
-    cloudwatch_client = get_cloudwatch_client()
-    current_timestamp = datetime.datetime.utcnow()
-    try:
-        response = cloudwatch_client.put_metric_data(
-            Namespace=namespace,
-            MetricData=[
-                {
-                    "MetricName": metric_name,
-                    "Dimensions": dimensions,
-                    "Value": value,
-                    "Unit": unit,
-                    "Timestamp": current_timestamp,
-                }
-            ],
-        )
-    except ClientError as e:
-        LOGGER.error("Error: Cannot put data to cloudwatch metric: {}".format(metric_name))
-        LOGGER.error("Exception: {}".format(e))
-        raise e
-
-def read_metric(csv_file):
-    df = pd.read_csv(csv_file)
-    value = df[df.columns[-1]].iloc[0]
-    if isinstance(value, str):
-        for i in range(len(value)):
-            if value[i].isdecimal() or value[i] == ".":
-                continue
-            else:
-                return float(value[:i])
-    return float(value)
