+++ test/sagemaker_tests/pytorch/training/integration/sagemaker/test_performance_inductor.py
@@ -26,7 +26,7 @@ from packaging.specifiers import SpecifierSet
 from ...integration import training_dir, DEFAULT_TIMEOUT, inductor_path
 from ...integration.sagemaker.timeout import timeout
 from .... import invoke_pytorch_helper_function
-from .test_inductor_helper import put_metric_data, read_metric
+# from .test_inductor_helper import put_metric_data, read_metric
 
 instance_types=("ml.p3.2xlarge", "ml.g5.4xlarge", "ml.g4dn.4xlarge")
 
@@ -57,7 +57,6 @@ def test_inductor_timm(framework_version, ecr_image, sagemaker_regions, instance
             'suites': "timm",
             'tmpdir': tmpdir,
         }
-
     invoke_pytorch_helper_function(ecr_image, sagemaker_regions, _test_inductor_performance, function_args)
 
 @pytest.mark.processor("gpu")
@@ -76,40 +75,53 @@ def test_inductor_torchbench(framework_version, ecr_image, sagemaker_regions, in
     invoke_pytorch_helper_function(ecr_image, sagemaker_regions, _test_inductor_performance, function_args)
 
 def _test_inductor_performance(ecr_image, sagemaker_session, framework_version, instance_type, suites, tmpdir):
-    with timeout(minutes=DEFAULT_TIMEOUT):   
-        output_path = f"s3://sagemaker-inductor-test/{suites}"
-        pytorch = PyTorch(
-            entry_point=f"run_inductor_{suites}.sh",
-            source_dir=inductor_path,
-            role='SageMakerRole',
-            instance_count=1,
-            instance_type=instance_type,
-            sagemaker_session=sagemaker_session,
-            image_uri=ecr_image,
-            framework_version=framework_version,
-            output_path=output_path,
-        )
-        pytorch.fit(job_name=utils.unique_name_from_base(f'test-pt-performance-inductor-{suites}'))
-        dimensions = [
-             {"Name": "InstanceType", "Value": instance_type},
-             {"Name": "ModelSuite", "Value": suites},
-             {"Name": "Precision", "Value": "AMP"},
-             {"Name": "WorkLoad", "Value": "Training"},
-        ]
-        s3_artifact_path = os.path.join(output_path,job_name,"output","output.tar.gz")
-        local_artifact = os.path.join(tmpdir, "output.tar.gz")
-        subprocess.check_output(["aws", "s3", "cp", s3_artifact_path, local_artifact])
-        with tarfile.open(local_artifact, "r:gz") as result:
-            result.extractall(path=tmpdir)
-        result_path = os.path.join(tmpdir, "benchmark", "bin","pytorch",f"{suites}_logs")
-        speedup = read_metric(os.path.join(result_path, "geomean.csv"))
-        comp_time = read_metric(os.path.join(result_path,"comp_time.csv"))
-        memory = read_metric(os.path.join(result_path,"memory.csv"))
-        passrate = read_metric(os.path.join(result_path,"passrate.csv"))
-        namespace = "PyTorch/SM/Benchmarks/TorchDynamo/Inductor"
-        put_metric_data("Speedup", namespace, "None", speedup, dimensions)
-        put_metric_data("CompilationTime", namespace, "Seconds", comp_time, dimensions)
-        put_metric_data("PeakMemoryFootprintCompressionRatio", namespace, "None", memory, dimensions)
-        put_metric_data("PassRate", namespace, "Percent", passrate, dimensions)
-
+    # with timeout(minutes=DEFAULT_TIMEOUT):   
+    output_path = f"s3://sagemaker-inductor-test/{suites}"
+    # job_name = "test-pt-performance-inductor-huggingface-1680825282-71e4"
+    # pytorch = PyTorch(
+    #     entry_point="test_inductor_helper.py",
+    #     source_dir=inductor_path,
+    #     role='SageMakerRole',
+    #     instance_count=1,
+    #     instance_type=instance_type,
+    #     sagemaker_session=sagemaker_session,
+    #     image_uri=ecr_image,
+    #     framework_version=framework_version,
+    #     hyperparameters = {'output_path':output_path,
+    #                         'job_name':job_name,
+    #                         'instance_type':instance_type,
+    #                         # 'tmpdir': tmpdir,
+    #                         'suites':suites
+    #                     },
+    # )
+    # pytorch.fit(job_name=utils.unique_name_from_base(f'upload-pt-performance-inductor-{suites}'))
+    pytorch = PyTorch(
+        entry_point=f"run_inductor_{suites}.sh",
+        source_dir=inductor_path,
+        role='SageMakerRole',
+        instance_count=1,
+        instance_type=instance_type,
+        sagemaker_session=sagemaker_session,
+        image_uri=ecr_image,
+        framework_version=framework_version,
+        output_path=output_path,
+    )
+    pytorch.fit(job_name=utils.unique_name_from_base(f'test-pt-performance-inductor-{suites}'))
+    job_name = pytorch.latest_training_job.name
+    pytorch = PyTorch(
+        entry_point="test_inductor_helper.py",
+        source_dir=inductor_path,
+        role='SageMakerRole',
+        instance_count=1,
+        instance_type=instance_type,
+        sagemaker_session=sagemaker_session,
+        image_uri=ecr_image,
+        framework_version=framework_version,
+        hyperparameters = {'output_path':output_path,
+                            'job_name':job_name,
+                            'instance_type':instance_type,
+                            'suites':suites,
+                        },
+    )
+    pytorch.fit(job_name=utils.unique_name_from_base(f'upload-pt-performance-inductor-{suites}'))
